+++ src/defaults.ts
@@ -3,6 +3,7 @@ export const DEFAULTS: PopcornDefaults = {
   seatColor: 'lightgrey',
   bookedColor: 'red',
   selectedColor: '#00356D',
+  unavailableColor: 'black',
   seatWidth: 30,
   seatMargin: 15,
   rowLabelWidth: 25,