+++ src/shapes/availableSeat.ts
@@ -1,5 +1,0 @@
-import SeatShape from './seatShape';
-
-class AvailableSeat extends SeatShape {}
-
-export default AvailableSeat;