+++ src/shapes/bookedSeat.ts
@@ -1,5 +1,0 @@
-import SeatShape from './seatShape';
-
-class BookedSeat extends SeatShape {}
-
-export default BookedSeat;