+++ src/shapes/selectedSeat.ts
@@ -1,5 +1,0 @@
-import SeatShape from './seatShape';
-
-class SelectedSeat extends SeatShape {}
-
-export default SelectedSeat;